Fixing requestFinish computation. Fixing some unit tests.

+++ src/test/java/provisioning/RanjanProvisioningSystemForHeterogeneousMachinesTest.java
@@ -64,7 +64,7 @@ public class RanjanProvisioningSystemForHeterogeneousMachinesTest extends Mocked
 		double averageUtilisation = 0.7;
 		long totalRequestsArrivals = 10;
 		long totalRequestsCompletions = 10;
-		long totalNumberOfServers = 1;
+		int totalNumberOfServers = 1;
 		MachineStatistics statistics = new MachineStatistics(averageUtilisation, totalRequestsArrivals, totalRequestsCompletions, totalNumberOfServers);
 		
 		//Mocks
@@ -92,7 +92,7 @@ public class RanjanProvisioningSystemForHeterogeneousMachinesTest extends Mocked
 		double averageUtilisation = 1.0;
 		long totalRequestsArrivals = 55;
 		long totalRequestsCompletions = 15;
-		long totalNumberOfServers = 3;
+		int totalNumberOfServers = 3;
 		MachineStatistics statistics = new MachineStatistics(averageUtilisation, totalRequestsArrivals, totalRequestsCompletions, totalNumberOfServers);
 		
 		//Mocks
@@ -120,7 +120,7 @@ public class RanjanProvisioningSystemForHeterogeneousMachinesTest extends Mocked
 		double averageUtilisation = 0.3;
 		long totalRequestsArrivals = 333;
 		long totalRequestsCompletions = 279;
-		long totalNumberOfServers = 20;
+		int totalNumberOfServers = 20;
 		MachineStatistics statistics = new MachineStatistics(averageUtilisation, totalRequestsArrivals, totalRequestsCompletions, totalNumberOfServers);
 		
 		//Mocks
@@ -148,7 +148,7 @@ public class RanjanProvisioningSystemForHeterogeneousMachinesTest extends Mocked
 		double averageUtilisation = 0.01;
 		long totalRequestsArrivals = 100;
 		long totalRequestsCompletions = 100;
-		long totalNumberOfServers = 20;
+		int totalNumberOfServers = 20;
 		MachineStatistics statistics = new MachineStatistics(averageUtilisation, totalRequestsArrivals, totalRequestsCompletions, totalNumberOfServers);
 		
 		//Mocks
@@ -174,7 +174,7 @@ public class RanjanProvisioningSystemForHeterogeneousMachinesTest extends Mocked
 		double averageUtilisation = 0.0;
 		long totalRequestsArrivals = 0;
 		long totalRequestsCompletions = 0;
-		long totalNumberOfServers = 20;
+		int totalNumberOfServers = 20;
 		MachineStatistics statistics = new MachineStatistics(averageUtilisation, totalRequestsArrivals, totalRequestsCompletions, totalNumberOfServers);
 		
 		//Mocks
@@ -202,7 +202,7 @@ public class RanjanProvisioningSystemForHeterogeneousMachinesTest extends Mocked
 		double averageUtilisation = 0.0;
 		long totalRequestsArrivals = 100;
 		long totalRequestsCompletions = 0;
-		long totalNumberOfServers = 0;
+		int totalNumberOfServers = 0;
 		MachineStatistics statistics = new MachineStatistics(averageUtilisation, totalRequestsArrivals, totalRequestsCompletions, totalNumberOfServers);
 		
 		//Mocks
@@ -235,7 +235,7 @@ public class RanjanProvisioningSystemForHeterogeneousMachinesTest extends Mocked
 		double averageUtilisation = 0.0;
 		long totalRequestsArrivals = 100;
 		long totalRequestsCompletions = 0;
-		long totalNumberOfServers = 0;
+		int totalNumberOfServers = 0;
 		MachineStatistics statistics = new MachineStatistics(averageUtilisation, totalRequestsArrivals, totalRequestsCompletions, totalNumberOfServers);
 		
 		JEEventScheduler scheduler = EasyMock.createStrictMock(JEEventScheduler.class);
@@ -303,7 +303,7 @@ public class RanjanProvisioningSystemForHeterogeneousMachinesTest extends Mocked
 		double averageUtilisation = 1.0;
 		long totalRequestsArrivals = 55;
 		long totalRequestsCompletions = 15;
-		long totalNumberOfServers = 3;
+		int totalNumberOfServers = 3;
 
 		MachineStatistics statistics = new MachineStatistics(averageUtilisation, totalRequestsArrivals, totalRequestsCompletions, totalNumberOfServers);
 		
@@ -390,7 +390,7 @@ public class RanjanProvisioningSystemForHeterogeneousMachinesTest extends Mocked
 		double averageUtilisation = 1.0;
 		long totalRequestsArrivals = 55;
 		long totalRequestsCompletions = 15;
-		long totalNumberOfServers = 3;
+		int totalNumberOfServers = 3;
 
 		MachineStatistics statistics = new MachineStatistics(averageUtilisation, totalRequestsArrivals, totalRequestsCompletions, totalNumberOfServers);
 		
@@ -465,7 +465,7 @@ public class RanjanProvisioningSystemForHeterogeneousMachinesTest extends Mocked
 		double averageUtilisation = 1.0;
 		long totalRequestsArrivals = 55;
 		long totalRequestsCompletions = 15;
-		long totalNumberOfServers = 3;
+		int totalNumberOfServers = 3;
 
 		MachineStatistics statistics = new MachineStatistics(averageUtilisation, totalRequestsArrivals, totalRequestsCompletions, totalNumberOfServers);
 		
@@ -540,7 +540,7 @@ public class RanjanProvisioningSystemForHeterogeneousMachinesTest extends Mocked
 		double totalUtilization = 0.2;
 		long totalRequestsArrivals = 100;
 		long totalRequestsCompletions = 100;
-		long totalNumberOfServers = 20;
+		int totalNumberOfServers = 20;
 		MachineStatistics statistics = new MachineStatistics(totalUtilization, totalRequestsArrivals, totalRequestsCompletions, totalNumberOfServers);
 		
 		SimpleSimulator configurable = EasyMock.createMock(SimpleSimulator.class);
